cbc-guidance-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==3.2.2
langchain==0.1.20
langchain-core==0.1.52
langchain-pinecone==0.1.0
langchain-huggingface==0.0.1
sentence-transformers==2.6.1
transformers==4.41.2
numpy==1.26.4
psycopg2-binary==2.9.9
google-generativeai==0.8.6
pypdf==4.0.1
python-docx==0.8.11
pydantic==2.7.4
